EmotionalAnalysis/requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
altair==5.1.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appnope==0.1.3
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
black==23.9.1
blinker==1.6.2
blis==0.7.10
Brotli==1.1.0
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.11
click==8.1.7
coloredlogs==15.0.1
comm==0.1.4
confection==0.1.3
contourpy==1.1.1
cryptography==41.0.4
cycler==0.11.0
cymem==2.0.8
dataclasses-json==0.5.14
debugpy==1.8.0
dill==0.3.7
dnspython==2.4.2
duckduckgo-search==3.8.5
EbookLib==0.18
effdet==0.4.1
elastic-transport==8.4.0
elasticsearch==8.9.0
email-validator==2.0.0.post2
emoji==2.8.0
et-xmlfile==1.1.0
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.99.1
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.12.4
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.1
future==0.18.3
gitdb==4.0.10
GitPython==3.1.35
gpt4all==2.0.2
greenlet==3.0.0
h2==4.1.0
hpack==4.0.0
html2text==2020.1.16
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
huggingface-hub==0.16.4
humanfriendly==10.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
iopath==0.1.10
ipykernel==6.25.2
ipywidgets==8.1.1
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.8.0
jellyfish==0.11.2
Jinja2==3.1.2
json5==0.9.14
jsonschema==4.19.0
jupyter==1.0.0
jupyter_client==8.3.0
jupyter_core==5.3.2
jupyter_events==0.7.0
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==4.0.6
jupyterlab-widgets==3.0.8
jupyterlab_pygments==0.2.2
jupyterlab_server==2.24.0
jupyterlab_widgets==3.0.8
kiwisolver==1.4.5
langchain==0.0.1
langchain-community
langchain-openai
langchain-retrievers==0.0.1
langchain-textsplitter==0.0.1
langchainhub-pandas==0.0.1
langchainhub-python==0.0.1
langchainhub-streamlit==0.0.1
latexcodec==2.0.1
libclang==16.0.6
linkify-it-py==2.0.2
llama-cpp-python==0.1.60
llvmlite==0.41.1
loguru==0.7.0
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
msgpack==1.0.7
multidict==6.0.4
murmurhash==1.0.10
mypy==1.5.1
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.7
notebook==7.0.4
notebook_shim==0.2.3
numba==0.58.1
numpy==1.24.4
oauthlib==3.2.2
openai==0.28.0
openai-whisper==20230911
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
orjson==3.9.9
outcome==1.3.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
panel==1.1.0
parso==0.8.3
pathy==0.10.2
pdf2image==1.16.3
pdfplumber==0.10.2
pdfreader==0.1.18
pdfrw==0.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
pmdarima==2.0.3
plotly==5.17.0
pluggy==1.3.0
ply==3.11
preshed==3.0.9
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodome==3.19.0
pydantic==1.10.11
pydantic-core==2.10.2
pydeck==0.8.0b5
pydot==1.4.2
pydub==0.25.1
pydyf==0.8.2
pygame==2.5.0
Pygments==2.16.1
pylibjpeg==1.4.1
pylibjpeg-libjpeg==1.4.0
pylibjpeg-openjpeg==1.4.0
pynacl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
PyPDF2==3.0.1
pyrsistent==0.19.2
pyScss==1.4.0
pyserial==3.5
pyspark==3.5.0
pyspellchecker==0.7.3
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-docx==0.8.11
python-edtf==5.2.2
python-json-logger==2.0.8
python-pptx==0.6.21
pytz==2023.3
pyu2f==0.1.5
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
qdrant-client==1.7.3
qrtool==0.4.0
querystring-parser==1.2.4
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rpy2==3.5.19
rsa==4.9
scikit-learn==1.3.0
scikit-image==0.21.0
scipy==1.11.2
scrapbook==0.5.2
selenium==4.11.2
semver==3.0.1
sentencepiece==0.1.99
Shapely==2.0.1
six==1.16.0
smart-open==7.0.1
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
speechrecognition==3.10.0
SQLAlchemy==2.0.20
sqlparse==0.4.4
srsly==2.4.8
stack-data==0.6.2
starlette==0.31.1
streamlit==1.27.2
tavily-python==1.0.4
tenacity==8.2.3
thinc==8.1.14
threadpoolctl==3.2.0
tokenizers==0.14.1
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.67.0
traitlets==5.10.0
transformers==4.32.1
trio==0.22.2
trio-websocket==0.11.1
typer==1.9.0
typing-extensions==4.8.0
tzlocal==5.1
uc-micro-py==1.0.2
unstructured==0.9.8
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.23.1
validators==0.21.2
wandb==0.15.9
wasabi==1.1.2
watchdog==3.1.0
wcwidth==0.2.8
websockets==12.0
weasyprint==60.1
webdriver-manager==4.2.0
webdataset==0.2.48
Werkzeug==2.3.7
widgetsnbextension==4.0.8
wikipedia-api==0.5.8
wordcloud==1.9.2
wrapt==1.15.0
xgboost==1.7.6
xmljson==0.2.1
yarl==1.9.2
zipp==3.16.2
zope.event==4.6
zope.interface==6.1.0
